Metaprogram: dictionary index commonalities.
